Add seed data

+++ backend/migrations/20170429032034-create-recipe-item.js
@@ -20,7 +20,7 @@ module.exports = {
       },
       quantity: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DOUBLE
       },
       unit: {
         allowNull: false,
